requirements.txt
fastmcp>=0.1.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
uvicorn>=0.20.0
websockets>=10.0
chinesecalendar>=1.9.0
holidays>=0.40
